requirements.txt
python-telegram-bot==20.6
docx2pdf
PyMuPDF
python-docx
pypandoc
